Update ch-5.4.html
Rationale for Change:

This issue was submitted via the IT Infrastructure Comment Form:

Table 5.4.3-8 specifies "ht:tComment" as the datatype for the "comments" element. While the datatype for e.g. "input" and "output" ("ht:tMessagePartsData") can contain a list, this datatype seems to be for one comment only, contradicting to the name and description of the element.

The datatype for the "comment" element should be a type that allows multiple comments, probably "ht:tComments".

GC: Looking at XSD files the element taskData/comments has the ht:tComments datatype not ht:tComment so the table 5.4.3-8 in volume 3 shall be updated.